requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0

